NU-1080 format files, and create folder for them
