+++ designer/client/src/components/graph/Graph.tsx
@@ -4,7 +4,7 @@ import { cloneDeep, debounce, isEmpty, isEqual, keys, sortBy, without } from "lo
 import React from "react";
 import { findDOMNode } from "react-dom";
 import "../../stylesheets/graph.styl";
-import { filterDragHovered, getLinkNodes, setLinksHovered } from "./dragHelpers";
+import { filterDragHovered, getLinkNodes, setLinksHovered } from "./utils/dragHelpers";
 import { updateNodeCounts } from "./EspNode/element";
 import { GraphPaperContainer } from "./focusable";
 import { applyCellChanges, calcLayout, createPaper, isModelElement } from "./GraphPartialsInTS";
@@ -15,14 +15,14 @@ import { PanZoomPlugin } from "./PanZoomPlugin";
 import { RangeSelectedEventData, RangeSelectPlugin, SelectionMode } from "./RangeSelectPlugin";
 import "./svg-export/export.styl";
 import { prepareSvg } from "./svg-export/prepareSvg";
-import * as GraphUtils from "./GraphUtils";
+import * as GraphUtils from "./utils/graphUtils";
 import { ComponentDragPreview } from "../ComponentDragPreview";
 import { rafThrottle } from "./rafThrottle";
 import { isEdgeEditable } from "../../common/EdgeUtils";
 import { NodeId, NodeType, Process, ProcessDefinitionData } from "../../types";
 import { Layout, NodePosition, Position } from "../../actions/nk";
 import { UserSettings } from "../../reducers/userSettings";
-import { GraphProps } from "./GraphWrapped";
+import { GraphProps } from "./types/index";
 import User from "../../common/models/User";
 import { updateLayout } from "./GraphPartialsInTS/updateLayout";
 import { getDefaultLinkCreator } from "./EspNode/link";
